+++ doc/source/changelog.rst
@@ -7,6 +7,8 @@ Next Version
 ------------
 
   * Fix memoryview construction from symbolic bytes.
+  * Fix error constructing int.from_bytes with an unsized iterable argument.
+    (fixes `#291 <https://github.com/pschanely/CrossHair/issues/291>`__)
 
 
 Version 0.0.65